♻️ F11: ORDERPAGE > Create customs hooks : useSuccessMessage and useMenu
