+++ src/components/pages/order/Main/Panel/FormAdmin/AddForm.js
@@ -1,26 +1,19 @@
-import { useState, useContext } from "react";
+import { useContext } from "react";
 import PanelContext from "../../../../../context/OrderContext";
 import { EMPTY_PRODUCT } from "./../../../../../../enums/product";
 import Form from "./Form";
 import SubmitMessage from "./SubmitMessage";
 import Button from "./../../../../../reusable-ui/Button";
+import { useMessageSuccess } from "./../../../../../../hooks/useSuccessMessage";
 
 export default function AddForm() {
   //this.state.
   const { handleAdd, newProduct, setNewProduct } = useContext(PanelContext);
-  const [isSubmitted, setIsSubmitted] = useState(false);
-
+  const { isSubmitted, displayMessageSuccess } = useMessageSuccess();
   //comportement
   const handleChange = (e) => {
     const { name, value } = e.target;
     setNewProduct({ ...newProduct, [name]: value });
-  };
-
-  const displayMessageSuccess = () => {
-    setIsSubmitted(true);
-    setTimeout(() => {
-      setIsSubmitted(false);
-    }, 2000);
   };
 
   const handleSubmit = (e) => {
