+++ src/components/pages/order/OrderPage.js
@@ -4,9 +4,8 @@ import Navbar from "../order/Navbar/Navbar";
 import Main from "../order/Main/Main";
 import PanelContext from "../../context/OrderContext";
 import { theme } from "./../../../theme/index";
-import { fakeMenu } from "../../api/fakeData/fakeMenu";
 import { EMPTY_PRODUCT } from "./../../../enums/product";
-// import { cloneArray } from "./../../../utils/array"; //@FIXME
+import { useMenu } from "../../../hooks/useMenu";
 
 export default function OrderPage() {
   //state
@@ -14,53 +13,52 @@ export default function OrderPage() {
   const [tabIconSelected, setTabIconSelected] = useState(false);
   const [tabSelected, setTabSelected] = useState("add");
   const [isCollasped, setIsCollasped] = useState(false);
-  const [menu, setMenu] = useState(fakeMenu.MEDIUM);
   const [newProduct, setNewProduct] = useState(EMPTY_PRODUCT);
   const [productSelected, setproductSelected] = useState(EMPTY_PRODUCT);
   const titleEditRef = useRef();
   const [hasAlreadyBeenClicked, setHasAlreadyBeenClicked] = useState(false);
-
+  const { menu, handleAdd, handleEdit, handleDelete, resetMenu } = useMenu();
   //comportement
   useEffect(() => {
     document.title = "Crazee Burger | Menu";
   }, []);
 
-  const handleAdd = (newProduct) => {
-    // copy du array
-    const menuCopy = [...menu];
-    //2 manip copy array
-    const menuUpdated = [newProduct, ...menuCopy];
-    //3 update du state
-    setMenu(menuUpdated);
-  };
+  // const handleAdd = (newProduct) => {
+  //   // copy du array
+  //   const menuCopy = [...menu];
+  //   //2 manip copy array
+  //   const menuUpdated = [newProduct, ...menuCopy];
+  //   //3 update du state
+  //   setMenu(menuUpdated);
+  // };
 
-  const handleDelete = (idOfProductToDelete) => {
-    //1 copy this.state.
-    const menuCopy = JSON.parse(JSON.stringify(menu));
-    //2 manip copy this.state.
-    const menuUpdated = menuCopy.filter(
-      (product) => product.id !== idOfProductToDelete
-    );
-    //3 update this.state.
-    setMenu(menuUpdated);
-  };
+  // const handleDelete = (idOfProductToDelete) => {
+  //   //1 copy this.state.
+  //   const menuCopy = JSON.parse(JSON.stringify(menu));
+  //   //2 manip copy this.state.
+  //   const menuUpdated = menuCopy.filter(
+  //     (product) => product.id !== idOfProductToDelete
+  //   );
+  //   //3 update this.state.
+  //   setMenu(menuUpdated);
+  // };
 
-  const handleEdit = (productBeingEdited) => {
-    // copy du array
-    const menuCopy = JSON.parse(JSON.stringify(menu));
-    //2 manip copy array
-    const indexOfProductToEdit = menu.findIndex(
-      (product) => product.id === productBeingEdited.id
-    );
-    menuCopy[indexOfProductToEdit] = productBeingEdited;
-    //3 update du state
-    setMenu(menuCopy);
-  };
+  // const handleEdit = (productBeingEdited) => {
+  //   // copy du array
+  //   const menuCopy = JSON.parse(JSON.stringify(menu));
+  //   //2 manip copy array
+  //   const indexOfProductToEdit = menu.findIndex(
+  //     (product) => product.id === productBeingEdited.id
+  //   );
+  //   menuCopy[indexOfProductToEdit] = productBeingEdited;
+  //   //3 update du state
+  //   setMenu(menuCopy);
+  // };
 
-  const resetMenu = () => {
-    setMenu(fakeMenu.SMALL);
-    setHasAlreadyBeenClicked(false);
-  };
+  // const resetMenu = () => {
+  //   setMenu(fakeMenu.SMALL);
+  //   setHasAlreadyBeenClicked(false);
+  // };
 
   const panelContextValue = {
     isModeAdmin,
